refactor: move SetUpCustomCommands to utils.go and update command constants
- Moved the SetUpCustomCommands function from stage_r1.go to utils.go for better organization.
- Updated command constants to use 'ls' and 'cat' instead of 'cls' and 'ccat'.
- Enhanced the command setup logic to improve clarity and maintainability.
